+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.0.0 - 2015-04-30
+
+* [#4](https://github.com/atoum/reports-extension/pull/4) Add telemetry report ([@jubianchi])
+
+## BC break updates
+
+* [#4](https://github.com/atoum/reports-extension/pull/4) PHP `< 5.5.0` is not supported anymore ([@jubianchi])
+
 # 1.0.1 - 2015-09-16
 
 * Loosened constraint on symfony/filesystem
@@ -5,3 +13,5 @@
 # 1.0.0 - 2015-05-13
 
 * HTML Coverage report with branches and paths support
+
+[@jubianchi]: https://github.com/jubianchi